When selecting a regiment on the regiment selection page, it will always be replaced by a new instance.
Refactored the regiment service to make use of promises. All regiments are now no longer openly accessible to service users, only names and the current selected one.

Added a modal that warns if changes would be discarded by choosing a new regiment and allows cancelling the change to keep them.