Changed the implementation of the kit choices.
Now, instead of each effect being selected by the user and it's effects applied one at a tme, the user makes all the choices and then all the effects are applied.

This makes it easier to reapply and remove them and allows cancelling inbetween by the user without having to try to roll anything back.