Removed the duplicate placeholder replacement logic in characteroptions and regimentoptions and replaced it with a single service.
Moved the regiments and specialties that a character can choose out of characteroptions and into individual services due to both relying on the placeholder service. In characteroptions this caused a circular dependency as it required injecting the placeholder service and the placeholder required characteroptions to get the objects it replaces the placeholders with.