Reorganized the Skill type.
Skills are now described by a SkillDescription class, which contains the name and aptitudes of the skill and specialization, if it has one.

The Skill class proper now represents a skill as it exists on a character and includes a SkillDescription and a collection of CharacterModifiers which change its rank.