Renamed characteroptions to characterOptions everywhere.
Fixed a bad url for the modal template in RegimentCreationElementController.

Changed the filter for regiment creation options to both consider point cost and if the option is already selected, so that duplicate doctrines can't be selected.

Reworked all the services to be more synchronous-like.

Deleted placeholders.js, it is superceded by PlaceholderReplacement.ts.

Renamed several json files.

Shuffled some npm dependencies around.